Add flow.Task interface
